backend/requirements.txt
# Flask and Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.4.4
Werkzeug==2.3.6
Flask-SocketIO==5.3.4
matlab
# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.20

# Environment and Configuration
python-dotenv==1.0.0

# Data Processing and Analysis
numpy==1.24.3
pandas==2.0.3
scipy==1.11.2

# Machine Learning and AI
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1
shap==0.42.1

# Image Processing
opencv-python==4.8.0.76
Pillow==10.0.0
matplotlib==3.7.2

# Geospatial Data
shapely==2.0.1
geopandas==0.13.2
folium==0.14.0

# Report Generation
reportlab==4.0.4
openpyxl==3.1.2

# Firebase and Notifications
firebase-admin==6.2.0
pyfcm==1.5.4

# API and HTTP
requests==2.31.0
urllib3==1.26.16

# Development and Testing
pytest==7.4.0
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0

# MATLAB Integration (Python Engine)
# Note: MATLAB Python Engine needs to be installed separately from MATLAB installation
# matlab==0.1  # Placeholder - actual installation requires MATLAB

# Utilities
python-dateutil==2.8.2
pytz==2023.3
celery==5.3.1
redis==4.6.0
